pyproject.toml
[tool.poetry]
name = "ticket-selling-platform"
version = "0.1.0"
description = "Selling tickets platfrom for vollybal games"
authors = ["Mohammad Taheri <admirer135@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
Django = "^4.2.7"
typing-extensions = "^4.8.0"
djangorestframework-simplejwt = "^5.3.0"
django-extensions = "^3.2.3"
drf-yasg = "^1.21.7"
django-rest-swagger = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
